ui-plain: redirect to add trailing slash for trees
If the user requests a directory but the URL does not contain a trailing
slash, issue a 301 redirect to add the trailing slash.  This fixes the
problem that, when a user requested a directory without a trailing
slash, all links in the directory listing were invalid.

Note that we do not include the query string in the redirect, since we
ignore the query string on this page.

Signed-off-by: Mark Lodato <lodatom@gmail.com>