ui-plain: add support for symbolic links
Follow symbolic links in the plain view.  A new tree-walking function
is needed, since read_tree_recursive() does not have an option to
dereference symbolic links within the directory portion of a match.
The new walk_tree() function operates on a single path and returns the
sha1 and mode of the resulting file.

Since a callback function is no longer needed, consolodate the
print_dir() family into a single function.

Write a simple normalize_path() to remove '.' and '' elements from
a path.  The existing normalize_path_copy() is not appropriate here,
since that function does not follow symbolic links.  For example, if
'x' links to directory 'a/b', then 'x/..' should return 'a', not ''.

Signed-off-by: Mark Lodato <lodatom@gmail.com>