ui-plain: add support for following symbolic links
Follow symbolic links in the plain view.  A new tree-walking function
is needed, since read_tree_recursive() does not have an option to
dereference symbolic links within the directory portion of a match.
The new cgit_find_object_by_path() function operates on a single path
and returns the sha1 and mode of the resulting file.  The new
'enable-symlink-traversal' option can be used to disable this behavior.

Since a callback function is no longer needed, consolidate the
print_dir() family into a single function.

cgit_find_object_by_path() needed a simple normalize_path() to remove
'.' and '' elements from a path.  The existing normalize_path_copy() is
not appropriate here, since that function does not follow symbolic
links.  For example, if 'x' links to directory 'a/b', then 'x/..' should
return 'a', not ''.

Signed-off-by: Mark Lodato <lodatom@gmail.com>