Refactor page titles
Give each cmd its own init function, in which page.title can be calculated
on a command-specific basis.  Many duplicate titles are avoided in this way.

By default, inject the command type and vpath breadcrumb into the title.
Specialise commit pages so that the the commit subject is in the title.

Remove the path breadcrumb from the page content as it is now in the title
anyway.